requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vanna[qdrant]==0.6.2
vanna[clickhouse]==0.6.2
python-dotenv==1.0.0
pydantic>=2.5.0
requests==2.31.0
qdrant-client>=1.7.0
clickhouse-connect>=0.6.0
pandas>=2.0.0
openai>=1.0.0
